backend/requirements.txt
certifi==2018.1.18
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
django-money==0.12.3
djangorestframework==3.7.7
pipenv==11.8.0
py-moneyed==0.7.0
pytz==2018.3
virtualenv==15.1.0
virtualenv-clone==0.3.0
